WIP: Move NodeRemoval to NodeAggregate subcontext
