+++ Neos.EventSourcedContentRepository/Classes/Domain/Context/NodeAggregate/Command/RemoveNodeFromAggregate.php
@@ -1,6 +1,6 @@
 <?php
 declare(strict_types=1);
-namespace Neos\EventSourcedContentRepository\Domain\Context\Node\Command;
+namespace Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Command;
 
 /*
  * This file is part of the Neos.ContentRepository package.
